Refactor app components for integration with router
