--- dronesym-frontend/src/app/app.component.css
@@ -1,9 +1,0 @@
-agm-map{
-	height: 100vh;
-}
-
-.control-buttons{
-	position: absolute;
-	bottom: 35px;
-	right: 35px;
-}
