+++ dronesym-frontend/src/app/dashboard/dashboard.component.css
@@ -1,0 +1,9 @@
+agm-map{
+	height: 100vh;
+}
+
+.control-buttons{
+	position: absolute;
+	bottom: 35px;
+	right: 35px;
+}
