+++ dronesym-frontend/src/app/dashboard/dashboard.component.html
@@ -1,0 +1,35 @@
+<agm-map #map [latitude]="centerCoords.lat" [longitude]="centerCoords.lon" [zoom] = "15"
+[streetViewControl]="false">
+	<agm-marker *ngFor="let waypoint of currDrone?.waypoints; let i = index" [latitude]="waypoint.lat"
+	[longitude]="waypoint.lon" (markerClick)="deleteWaypoint(i)" [iconUrl]="'assets/img/blue.png'"></agm-marker>
+
+	<agm-polyline *ngIf="currDrone?.waypoints.length > 1" strokeColor="blue" strokeOpacity="0.6">
+		<agm-polyline-point *ngFor="let waypoint of currDrone?.waypoints" [latitude]="waypoint.lat" [longitude]="waypoint.lon"> </agm-polyline-point>
+	</agm-polyline>
+
+	<agm-marker *ngFor="let i of droneIndices" [latitude]="drones[i].location.lat"
+		[longitude]="drones[i].location.lon" (markerClick)="setCurrentDrone(drones[i])">
+		<agm-info-window>
+			<drone-option-box (onSelected)="processDroneBox($event)"></drone-option-box>
+		</agm-info-window>
+	</agm-marker>
+
+</agm-map>
+
+<cursor-tooltip *ngIf="createMode == createModes.DRONES || createMode == createModes.WAYPOINTS" [latitude]="cursor.lat" [longitude]="cursor.lon" [x]="cursor.x" [y]="cursor.y"></cursor-tooltip>
+
+<confirm-dialog [show]="this.dialogParams.droneDialog.show" (onResponse)="processDialogResponse($event)"></confirm-dialog>
+
+<div class="control-buttons">
+	<div *ngIf="createMode == createModes.WAYPOINTS"  class="row">
+		<a (click)="finishAddingWaypoints()" class="btn-floating btn-large waves-effect waves-light green"><i class="material-icons">done</i></a>
+	</div>
+
+	<div *ngIf="createMode == createModes.WAYPOINTS"  class="row">
+		<a (click)="cancelAddingWaypoints()" class="btn-floating btn-large waves-effect waves-light red"><i class="material-icons">close</i></a>
+	</div>
+
+	<div *ngIf="createMode == createModes.NONE || createMode == createModes.DRONES" class="row">
+		<a (click)="goToCreateDroneMode()" class="btn-floating btn-large waves-effect waves-light blue"><i class="material-icons">add</i></a>
+	</div>
+</div>
