+++ dronesym-frontend/src/app/login/login.component.css
@@ -1,0 +1,8 @@
+.login-background{
+	background-color: #f5f5f5;
+}
+
+.login-card{
+	margin-top: 8%;
+	margin-bottom: 20%;
+}
