+++ dronesym-frontend/src/app/login/login.component.html
@@ -1,0 +1,25 @@
+<div class="row login-background">
+	<div class="col s4 offset-s4 login-card">
+		<div class="card">
+			<div class="card-content">
+				<span class="card-title">Login</span>
+
+				<form materialize>
+					<div class="input-field col s11">
+						<label for="username"> Username </label>
+						<input id="username" type="text"/>
+					</div>
+
+					<div class="input-field col s11">
+						<label for="password"> Password </label>
+						<input id="password" type="password"/>
+					</div>
+
+					<div class="row">
+						<button class="waves-effect waves-light btn blue darken-3 col s4 offset-s4">Login</button>
+					</div>
+				</form>
+			</div>
+		</div>
+	</div>
+</div>
